requirements.txt
pinecone-client>=3.0.0
transformers>=4.36.0
torch>=2.1.0
pymongo>=4.6.0
streamlit>=1.31.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
fastapi>=0.109.0
uvicorn>=0.27.0 